+++ projects/ngx-xchange-icons/src/lib/more-three-dots.component.ts
@@ -4,7 +4,7 @@ import { Component } from '@angular/core';
   selector: 'xc-icon-more-three-dots',
   standalone:true,
   template: `
-<svg xmlns="http://www.w3.org/2000/svg" width="100%" height="100%" viewBox="0 0 22 22" fill="none">
+<svg xmlns="http://www.w3.org/2000/svg"   viewBox="0 0 22 22"  fill="none" width="100%" height="100%">
 <path d="M11 18.3333C11.5062 18.3333 11.9166 17.9229 11.9166 17.4167C11.9166 16.9104 11.5062 16.5 11 16.5C10.4937 16.5 10.0833 16.9104 10.0833 17.4167C10.0833 17.9229 10.4937 18.3333 11 18.3333Z" stroke="#111111" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
 <path d="M11 11.9167C11.5062 11.9167 11.9166 11.5063 11.9166 11C11.9166 10.4937 11.5062 10.0833 11 10.0833C10.4937 10.0833 10.0833 10.4937 10.0833 11C10.0833 11.5063 10.4937 11.9167 11 11.9167Z" stroke="#111111" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>
 <path d="M11 5.50001C11.5062 5.50001 11.9166 5.0896 11.9166 4.58334C11.9166 4.07708 11.5062 3.66667 11 3.66667C10.4937 3.66667 10.0833 4.07708 10.0833 4.58334C10.0833 5.0896 10.4937 5.50001 11 5.50001Z" stroke="#111111" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"/>