+++ projects/ngx-xchange-icons/src/lib/navigator-board-selection.component.ts
@@ -1,0 +1,23 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'xc-icon-navigator-board-selection',
+  standalone:true,
+  template: `
+<svg xmlns="http://www.w3.org/2000/svg" width="100%" height="100%" viewBox="0 0 40 40" fill="none">
+<rect width="40" height="40" rx="10" fill="none" fill-opacity="0.5"/>
+<rect x="8" y="29" width="18" height="5" rx="2.5" transform="rotate(-90 8 29)" stroke="#111111" stroke-width="2" stroke-linecap="round"/>
+<rect x="17" y="29" width="18" height="5" rx="2.5" transform="rotate(-90 17 29)" stroke="#111111" stroke-width="2" stroke-linecap="round"/>
+<rect x="26" y="29" width="18" height="5" rx="2.5" transform="rotate(-90 26 29)" stroke="#111111" stroke-width="2" stroke-linecap="round"/>
+<defs>
+<linearGradient id="paint0_linear_3354_13388" x1="35.7547" y1="29.3548" x2="5.16067" y2="19.1152" gradientUnits="userSpaceOnUse">
+</linearGradient>
+</defs>
+</svg>
+`,
+  styles: [
+  ]
+})
+export class XcIconNavigatorBoardSelectionComponent {
+
+}