+++ projects/ngx-xchange-icons/src/public-api.ts
@@ -1143,3 +1143,8 @@ export * from './lib/more-three-dots.component';
 export * from './lib/close-xl.component';
 export * from './lib/close-s.component';
 export * from './lib/ai.component';
+export * from './lib/navigator-graph-selection.component';
+export * from './lib/navigator-table-selection.component';
+export * from './lib/navigator-board-selection.component';
+export * from './lib/success-icon.component'
+export * from './lib/no-assessments-image.component'