+++ projects/ngx-xchange-ui/src/components/check-rounded-button/check-rounded-button/check-rounded-button.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { CheckRoundedButtonComponent } from './check-rounded-button.component';
+
+describe('CheckRoundedButtonComponent', () => {
+  let component: CheckRoundedButtonComponent;
+  let fixture: ComponentFixture<CheckRoundedButtonComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [CheckRoundedButtonComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(CheckRoundedButtonComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});