+++ projects/ngx-xchange-ui/src/components/complete-objective-user-card/complete-objective-user-card/complete-objective-user-card.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
+import { Component, EventEmitter, input, Input, OnChanges, OnInit, Output, SimpleChanges } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 
 interface UserData {
@@ -21,22 +21,17 @@ interface UserData {
   templateUrl: './complete-objective-user-card.component.html',
   styleUrl: './complete-objective-user-card.component.css',
 })
-export class CompleteObjectiveUserCardComponent implements OnInit, OnChanges {
+export class CompleteObjectiveUserCardComponent implements OnInit {
   @Input() userData:UserData | undefined = undefined;
   @Output() checked: BehaviorSubject<boolean> = new BehaviorSubject<boolean>(false);
-  public readMode = false;
+  @Input() readMode:boolean = false;
   
   ngOnInit(): void {
-    this.readMode = this.userData?.isCompleted ? this.userData.isCompleted : false;
-  }
-  ngOnChanges(changes: SimpleChanges): void {
-    if(changes['userData']){
-      this.userData = changes['userData'].currentValue
-    }
   }
   emitChecked(){
-    this.checked.next(!this.userData?.isCompleted)
     if(this.userData)
       this.userData.isCompleted = !this.userData.isCompleted;
+    this.checked.next(Boolean(this.userData?.isCompleted))
   }
 }
+