+++ projects/ngx-xchange-ui/src/components/navigator-range-filter/navigator-range-component/navigator-range-filter.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { NavigatorRangeFilterComponent } from './navigator-range-filter.component';
+
+describe('NavigatorRangeFilterComponent', () => {
+  let component: NavigatorRangeFilterComponent;
+  let fixture: ComponentFixture<NavigatorRangeFilterComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [NavigatorRangeFilterComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(NavigatorRangeFilterComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});