+++ projects/ngx-xchange-ui/src/components/navigator-users-filter/navigator-users-filter/navigator-users-filter.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { NavigatorUsersFilterComponent } from './navigator-users-filter.component';
+
+describe('NavigatorUsersFilterComponent', () => {
+  let component: NavigatorUsersFilterComponent;
+  let fixture: ComponentFixture<NavigatorUsersFilterComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      imports: [NavigatorUsersFilterComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(NavigatorUsersFilterComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});