+++ projects/ngx-xchange-ui/src/components/objective-options/objective-options.module.ts
@@ -1,0 +1,26 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ObjectiveOptionsComponent } from './objective-options/objective-options.component';
+import { XcIconAddComponent, XcIconCopyComponent, XcIconEditComponent, XcIconFileShareAltComponent, XcIconTrashComponent } from '@indziaki/ngx-xchange-icons';
+import { MainHeadingModule } from '../../public-api';
+
+
+
+@NgModule({
+  declarations: [
+    ObjectiveOptionsComponent
+  ],
+  imports: [
+    CommonModule,
+    MainHeadingModule,
+    XcIconEditComponent,
+    XcIconCopyComponent,
+    XcIconFileShareAltComponent,
+    XcIconTrashComponent,
+    XcIconAddComponent
+  ],
+  exports: [
+    ObjectiveOptionsComponent
+  ]
+})
+export class ObjectiveOptionsModule { }