+++ projects/ngx-xchange-ui/src/directives/container/container.directive.ts
@@ -24,3 +24,4 @@ type containerTypes =
 |'listSelector'
 |'assignmentCard'
 |'rightSideModal'
+|'navigatorFilterContainer'